Deprecating Academy.Instance.FloatProperties (#3696)
* Deprecating Academy.Instance.FloatProperties

* Made the registered side channels a static property and created the sideChannelUtils class to handle side channel stuff

* Clearing the sending message queue in the Academy when the communicaor is not on

* addressing comments
