+++ com.unity.ml-agents/Runtime/SideChannels/SideChannel.cs
@@ -1,7 +1,5 @@
 using System.Collections.Generic;
 using System;
-using System.IO;
-using System.Text;
 
 namespace MLAgents.SideChannels
 {
